029-完成位置同步服务代码

+++ ConnectBridge/Vector3Data.cs
@@ -3,8 +3,8 @@ using System;
 namespace ConnectBridge {
     [Serializable]
     public class Vector3Data {
-        public float X { get; set; }
-        public float Y { get; set; }
-        public float Z { get; set; }
+        public double X { get; set; }
+        public double Y { get; set; }
+        public double Z { get; set; }
     }
 }
